o Made ZipArchiver aware of the 'includeEmptyDirs' setting (which kind of overrides   the 'filesOnly' setting; for now both 'work'.
o Made order in which the files/dirs are iterated predictable

o Reversed order of adding dirs/files in getFiles() so dirs are
  always specified first.

o Re-enable testcase where directory permissions are tested in ZipArchiver:
  Due to the non deterministic nature of 'addParentDirs' the last specified
  directoryMask was used for those files. Together with the abovementioned
  modifications the ZipArchiver now uses the default directory permissions
  in the correct time (effectively 'addParentDirs' doesn't have any real effect now..)

ZipArchiver is due for refactoring! I'm just not sure about the backward
compatibility requirements yet. If this isn't an issue it could get a lot
cleaner.


git-svn-id: http://svn.codehaus.org/plexus/trunk/plexus-components@2434 30dfa635-7bf8-0310-9608-e41f2d776689